// src/api/generated/model/examTypeDtoDepartment.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export type ExamTypeDtoDepartment =
  (typeof ExamTypeDtoDepartment)[keyof typeof ExamTypeDtoDepartment];

export const ExamTypeDtoDepartment = {
  TURKISH: "TURKISH",
  ENGLISH: "ENGLISH",
  GERMAN: "GERMAN",
  CHINESE: "CHINESE",
  ARABIC: "ARABIC",
  FRENCH: "FRENCH",
  JAPANESE: "JAPANESE",
  RUSSIAN: "RUSSIAN",
  KOREAN: "KOREAN",
  GREEK: "GREEK",
  PERSIAN: "PERSIAN",
} as const;