// src/api/generated/model/examTypeDtoStatus.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export type ExamTypeDtoStatus =
  (typeof ExamTypeDtoStatus)[keyof typeof ExamTypeDtoStatus];

export const ExamTypeDtoStatus = {
  ACTIVE: "ACTIVE",
  PASSIVE: "PASSIVE",
  DELETED: "DELETED",
} as const;